+++ server/quiz/quiz.go
@@ -17,7 +17,7 @@ type Question struct {
 	Choices []string `json:"choices"`
 	Answer  string   `json:"answer"`
 	Source  string   `json:"source"`
-	Colour  []uint8  `json:"colour"`
+	Colour  []int    `json:"colour"`
 }
 
 type Quiz struct {
